Refactor
Refactoring Encryptor from separate classes to one
